MP stances need to be grouped by id
a voter hash may vary over time, e.g. with different versions of their
name, so grouping by that will give us multiple stances for the same
person.

It may be better to live with that, and then combine them at the end,
but for now let’s go with collapsing any hash to its id (and cope later
with hashes that don’t have one…)