Split num_votes and num_motions
the PW scoring system gives points for abstaining … but the UK has no
differentiation between abstention and absence. So we end up with Sinn
Féin MPs being given a position based on abstentions.

We need a better way of handing this internationally, but for now
simply remove absences to get a num_votes that we can check at display
time (calculating the weight still uses num_motions, which includes any
motion whilst this person was an MP, and treats absence == abstention)